Box refactor
The Box's method `process` is a layer around the code itself.  It manages
required & provided values, captures and acts upon sentinel messages.
It receives a message, eventually pull the required data and passes them
into `processValue`.

The code to be customize is the method `processValue`.

The flow builder is relieved from messageData, required & provided fields.
It manipulates the whole Message leaving the details on the upper layer
of the Box class.