Reproducing test.
The problem occurs when there are several dimensions one being a way
faster than the other.  Then as the first dimension is complete, the
tracing tries to check the other dimension (it is inferred from the
dimension graph).  However, not even one message hasn't make it from
the first box of the other dimension thus having the parent message
dimension map without any info.